# docs/swagger.yaml
basePath: /api/calculator/v1
definitions:
  domain.Calculation:
    properties:
      a:
        type: number
      b:
        type: number
      operation:
        type: string
    type: object
  domain.CalculationResult:
    properties:
      result:
        type: string
    type: object
  domain.Record:
    properties:
      amount:
        type: number
      created_at:
        type: string
      id:
        type: string
      operation_id:
        type: string
      operation_response: {}
      user_balance:
        type: number
      user_id:
        type: string
    type: object
  domain.RecordSearch:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      search_term:
        type: string
      sort:
        type: string
      user_id:
        type: string
    type: object
  domain.User:
    properties:
      balance:
        type: number
      id:
        type: string
      status:
        $ref: '#/definitions/domain.UserStatus'
      username:
        type: string
    type: object
  domain.UserStatus:
    enum:
    - active
    - inactive
    type: string
    x-enum-varnames:
    - UserStatusActive
    - UserStatusInactive
  entities.BalanceResponse:
    properties:
      balance:
        type: number
    type: object
  entities.CreateUser:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  entities.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  entities.LoginResponse:
    properties:
      token:
        type: string
    type: object
  entities.LoginUser:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  entities.SearchRecordsResponse:
    properties:
      records:
        items:
          $ref: '#/definitions/domain.Record'
        type: array
    type: object
host: localhost:8080
info:
  contact: {}
  title: Calculator API
  version: "1.0"
paths:
  /balance:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.BalanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get user balance
      tags:
      - User
  /calculate:
    post:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.Calculation'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/domain.CalculationResult'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Perform a operation
      tags:
      - Calculator
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.LoginUser'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: User login
      tags:
      - User
  /records:
    post:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RecordSearch'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entities.SearchRecordsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Get paginated records by search term
      tags:
      - Records
  /records/:ID:
    delete:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/domain.RecordSearch'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/entities.SearchRecordsResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Delete record
      tags:
      - Records
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: query params
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entities.CreateUser'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/domain.User'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/entities.ErrorResponse'
      summary: Create a new user
      tags:
      - User
swagger: "2.0"